test: test_stl -> test_serialize
